+++ Structurizr.Core.Tests/Model/ContainerInstanceTests.cs
@@ -21,7 +21,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void Test_construction()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
     
             Assert.Same(_database, containerInstance.Container);
             Assert.Equal(_database.Id, containerInstance.ContainerId);
@@ -31,7 +31,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void Test_ContainerId()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
     
             Assert.Equal(_database.Id, containerInstance.ContainerId);
             containerInstance.Container = null;
@@ -42,7 +42,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void Test_Name()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
     
             Assert.Equal(_database.Name, containerInstance.Name);
     
@@ -52,21 +52,21 @@ namespace Structurizr.Core.Tests
     
         [Fact]
         public void test_CanonicalName() {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
     
             Assert.Equal("ContainerInstance://Default/Deployment Node/System.Database Schema[1]", containerInstance.CanonicalName);
         }
     
         [Fact]
         public void test_Parent_ReturnsTheParentDeploymentNode() {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
     
             Assert.Equal(_deploymentNode, containerInstance.Parent);
         }
     
         [Fact]
         public void test_RequiredTags() {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
     
             Assert.Equal(0, containerInstance.GetRequiredTags().Count);
         }
@@ -74,7 +74,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_getTags() {
             _database.AddTags("Database");
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
             containerInstance.AddTags("Primary Instance");
     
             Assert.Equal("Container Instance,Primary Instance", containerInstance.Tags);
@@ -83,7 +83,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_RemoveTags_DoesNotRemoveRequiredTags()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
     
             Assert.True(containerInstance.Tags.Contains(Tags.ContainerInstance));
     
@@ -95,7 +95,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
             Assert.Equal(0, containerInstance.HealthChecks.Count);
 
             HttpHealthCheck healthCheck = containerInstance.AddHealthCheck("Test web application is working", "http://localhost:8080");
@@ -109,7 +109,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck_ThrowsAnException_WhenTheNameIsNull()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
 
             try
             {
@@ -125,7 +125,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck_ThrowsAnException_WhenTheNameIsEmpty()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
 
             try
             {
@@ -141,7 +141,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck_ThrowsAnException_WhenTheUrlIsNull()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
 
             try
             {
@@ -157,7 +157,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck_ThrowsAnException_WhenTheUrlIsEmpty()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
 
             try
             {
@@ -173,7 +173,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck_ThrowsAnException_WhenTheUrlIsInvalid()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
 
             try
             {
@@ -189,7 +189,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck_ThrowsAnException_WhenTheIntervalIsLessThanZero()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
 
             try
             {
@@ -205,7 +205,7 @@ namespace Structurizr.Core.Tests
         [Fact]
         public void test_AddHealthCheck_ThrowsAnException_WhenTheTimeoutIsLessThanZero()
         {
-            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, true);
+            ContainerInstance containerInstance = Model.AddContainerInstance(_deploymentNode, _database, "Default");
 
             try
             {